Manually strip List<> for Target
Hackiest fix in existence.  Removes the problematic use of Match in relation to RegExp.  Now RegExp checks to esnure it is a list, then List<> is stripped and the resulting string is deserialized again.  A cumbersome and bulky way of doing this, but I don't want to stray too far from what works.